updater changes

+++ src/utils/updater.rs
@@ -8,9 +8,13 @@ pub fn local_update(new_bin: &str) -> i8 {
         return 0;
     }
 
+    if !Path::new("updated").exists() {
+        return -1;
+    }
+
     self_replace::self_replace(&new_bin).log_expect("Failed to update binary");
     std::fs::remove_file(&new_bin).log_expect("Failed to remove temp binary");
-    std::thread::sleep(std::time::Duration::from_secs(1));
+    std::fs::File::create("updated").log_expect("Failed to create temp file");
 
     if !Path::new("regy_bot_update_helper.exe").exists() {
         return 1;
